new canvases for first four levels

+++ put_border_on_canvas.m
@@ -14,7 +14,7 @@ else
 end
 
 
-new_img = zeros(rows+rows_to_add, cols+cols_to_add,3);
+new_img = uint8(zeros(rows+rows_to_add, cols+cols_to_add,3));
 new_img(rows_to_add+1:end, cols_to_add+1:end, :) = img;
 
 if(rows == 768)
